pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pydistort"
description = "Asynchronous media distortion module"
keywords = ["distort", "async"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dynamic = ["version", "readme"]
requires-python = ">=3.10"
dependencies  = [
    "pillow~=9.0",
]

[project.optional-dependencies]
apng = ["apng~=0.3.4"]
lottie = ["lottie[GIF]~=0.6.11"]
all = ["apng~=0.3.4", "lottie[GIF]~=0.6.11"]

[project.urls]
Homepage = "https://github.com/barsikus007/pydistort"
"Bug Tracker" = "https://github.com/barsikus007/pydistort/issues"

[[project.authors]]
name = "barsikus007"

[tool.setuptools.dynamic]
version = {attr = "pydistort.__version__"}
readme = {file = ["README.md", "LICENSE"]}
